Feat : Added layouts
